Dynamic load PDF viewer
It keeps it out of the tested code.
Jest can't load the PDF viewer.
